test(bmc-explorer): consolidate tests into integration suite
