+++ crates/bmc-explorer/tests/integration/bluefield3_explore.rs
@@ -14,12 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-mod common;
-
 use bmc_explorer::nv_generate_exploration_report;
 use bmc_mock::{DpuSettings, test_support};
 use model::site_explorer::EndpointType;
 use tokio::test;
+
+use crate::common;
 
 #[test]
 async fn explore_bluefield3_baseline() {
