+++ crates/bmc-explorer/tests/integration/dgx_gb300_explore.rs
@@ -14,36 +14,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-mod common;
-
+use bmc_explorer::hw::HwType;
 use bmc_explorer::nv_generate_exploration_report;
+use bmc_explorer::test_support::detect_hw_type;
 use bmc_mock::test_support;
 use model::site_explorer::EndpointType;
 use tokio::test;
 
+use crate::common;
+
+/// A DGX GB300 runs the NVIDIA "GB BMC" -- the same ServiceRoot signature as a
+/// Wiwynn GB200 (`Vendor: NVIDIA`, `Product: GB BMC`). It must classify as the
+/// GB300 platform (resolved from the NVIDIA GB300 GPU chassis ahead of the GB200
+/// arm), not as `Gb200`. `DgxGb300` maps to BMCVendor::Nvidia.
 #[test]
-async fn explore_liteon_power_shelf() {
-    let h = test_support::liteon_powershelf_bmc().await;
-    let report = nv_generate_exploration_report(h.service_root, &common::explorer_config())
+async fn explore_dgx_gb300() {
+    let h = test_support::dgx_gb300_bmc().await;
+    let config = common::explorer_config();
+
+    // Decisive assertion: a DGX GB300 must resolve to DgxGb300, not the Gb200
+    // fallback. Both map to BMCVendor::Nvidia, so asserting the report vendor
+    // alone would pass even with the DgxGb300 arm removed.
+    assert_eq!(
+        detect_hw_type(h.service_root.clone(), &config)
+            .await
+            .unwrap(),
+        Some(HwType::DgxGb300),
+    );
+
+    let report = nv_generate_exploration_report(h.service_root, &config)
         .await
         .unwrap();
-
     assert_eq!(report.endpoint_type, EndpointType::Bmc);
-    assert_eq!(report.vendor, Some(bmc_vendor::BMCVendor::Liteon));
+    assert_eq!(report.vendor, Some(bmc_vendor::BMCVendor::Nvidia));
     assert!(!report.systems.is_empty(), "systems must be present");
     assert!(!report.chassis.is_empty(), "chassis must be present");
-    assert!(
-        report
-            .service
-            .iter()
-            .any(|service| service.id == "FirmwareInventory"),
-        "firmware inventory service must be present"
-    );
-    assert!(
-        report
-            .machine_setup_status
-            .as_ref()
-            .is_some_and(|status| !status.diffs.is_empty() || status.is_done),
-        "machine setup status must be present and structurally valid"
-    );
 }
