+++ crates/bmc-explorer/tests/integration/hpe_proliant_dl380a_gen11_explore.rs
@@ -14,31 +14,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-mod common;
-
 use bmc_explorer::nv_generate_exploration_report;
 use bmc_mock::test_support;
-use model::site_explorer::EndpointType;
+use model::site_explorer::{EndpointType, InternalLockdownStatus};
 use tokio::test;
 
+use crate::common;
+
 #[test]
-async fn explore_wiwynn_gb200() {
-    let h = test_support::wiwynn_gb200_bmc().await;
+async fn explore_hpe_proliant_dl380a_gen11() {
+    let h = test_support::hpe_proliant_dl380a_gen11_bmc().await;
     let report = nv_generate_exploration_report(h.service_root, &common::explorer_config())
         .await
         .unwrap();
 
     assert_eq!(report.endpoint_type, EndpointType::Bmc);
-    assert_eq!(report.vendor, Some(bmc_vendor::BMCVendor::Nvidia));
+    assert_eq!(report.vendor, Some(bmc_vendor::BMCVendor::Hpe));
     assert!(!report.systems.is_empty(), "systems must be present");
     assert!(!report.chassis.is_empty(), "chassis must be present");
-    assert!(
-        report
-            .service
-            .iter()
-            .any(|service| service.id == "FirmwareInventory"),
-        "firmware inventory service must be present"
-    );
+
+    let system = &report.systems[0];
+    assert_eq!(system.model.as_deref(), Some("ProLiant DL380a Gen11"));
+
+    // The mock presents the locked-down production state:
+    // UsbBoot=Disabled and iLO VirtualNICEnabled=false.
+    let lockdown = report
+        .lockdown_status
+        .as_ref()
+        .expect("lockdown status must be present for HPE");
+    assert_eq!(lockdown.status, InternalLockdownStatus::Enabled);
+
     assert!(
         report
             .machine_setup_status
