+++ crates/bmc-explorer/tests/integration/main.rs
@@ -1,0 +1,31 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+mod bluefield3_explore;
+mod bluefield4_explore;
+mod common;
+mod dell_poweredge_r750_explore;
+mod dgx_gb300_explore;
+mod generic_ami;
+mod generic_supermicro_explore;
+mod hpe_proliant_dl380a_gen11_explore;
+mod lenovo_gb300_explore;
+mod nvidia_switch_explore;
+mod powershelf_explore;
+mod supermicro_gb300_explore;
+mod vera_rubin_explore;
+mod wiwynn_gb200_explore;
