--- crates/bmc-explorer/tests/supermicro_gb300_explore.rs
@@ -1,52 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-mod common;
-
-use bmc_explorer::hw::HwType;
-use bmc_explorer::nv_generate_exploration_report;
-use bmc_explorer::test_support::detect_hw_type;
-use bmc_mock::test_support;
-use model::site_explorer::EndpointType;
-use tokio::test;
-
-/// An SMC GB300 runs a Supermicro OpenBMC host BMC, so its ServiceRoot reports
-/// `Vendor: Supermicro` / `Product: GB NVL` with no OEM key. It must classify as
-/// the GB300 platform (resolved from the NVIDIA GB300 GPU chassis, then routed by
-/// the Supermicro vendor string), and `SupermicroGb300` maps to BMCVendor::Supermicro.
-#[test]
-async fn explore_supermicro_gb300() {
-    let h = test_support::supermicro_gb300_bmc().await;
-    let config = common::explorer_config();
-
-    // Decisive assertion: an SMC GB300 must resolve to SupermicroGb300, not the
-    // generic Supermicro fallback. Both map to BMCVendor::Supermicro, so a report
-    // vendor assertion alone would not prove the SupermicroGb300 arm was taken.
-    assert_eq!(
-        detect_hw_type(h.service_root.clone(), &config)
-            .await
-            .unwrap(),
-        Some(HwType::SupermicroGb300),
-    );
-
-    let report = nv_generate_exploration_report(h.service_root, &config)
-        .await
-        .unwrap();
-    assert_eq!(report.endpoint_type, EndpointType::Bmc);
-    assert_eq!(report.vendor, Some(bmc_vendor::BMCVendor::Supermicro));
-    assert!(!report.systems.is_empty(), "systems must be present");
-    assert!(!report.chassis.is_empty(), "chassis must be present");
-}
